requirements.txt
configobj==5.0.6
future==0.15.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-telegram-bot==3.4
six==1.10.0
Werkzeug==0.11.5
